+++ test/unit/quicknxs/interfaces/data_handling/test_data_manipulation.py
@@ -1,5 +1,3 @@
-# package imports
-# 3rd-party imports
 import copy
 
 import mantid.simpleapi as api
@@ -11,6 +9,7 @@ from quicknxs.interfaces.data_handling.data_manipulation import (
     NormalizeToUnityQCutoffError,
     _get_polynomial_fit_stitching_scaling_factor,
     _get_stitching_overlap_region,
+    extract_meta_data,
     generate_short_script,
     smart_stitch_reflectivity,
 )
@@ -27,15 +26,15 @@ mock_reduced_file_str = (
     "#\n"
     "# [Direct Beam Runs]\n"
     "#    DB_ID        P0        PN     x_pos   x_width     y_pos   y_width    bg_pos  bg_width      dpix       tth    number      File\n"  # noqa: E501
-    "#        1         0         0       195        12     126.5       155        55        24       194         0     42099  quicknxs-data/REF_M_42099.nxs.h5\n"  # noqa: E501
-    "#        2         0         0       195        12       127       154        55        24       194         0     42100  quicknxs-data/REF_M_42100.nxs.h5\n"  # noqa: E501
-    "#        3         0         0       195        12       127       154        55        24       194         0     42100  quicknxs-data/REF_M_42100.nxs.h5\n"  # noqa: E501
+    "#        1         0         0       195        12     126.5       155        55        24       194         0     42099  data/quicknxs-data/REF_M_42099.nxs.h5\n"  # noqa: E501
+    "#        2         0         0       195        12       127       154        55        24       194         0     42100  data/quicknxs-data/REF_M_42100.nxs.h5\n"  # noqa: E501
+    "#        3         0         0       195        12       127       154        55        24       194         0     42100  data/quicknxs-data/REF_M_42100.nxs.h5\n"  # noqa: E501
     "#\n"
     "# [Data Runs]\n"
     "#    scale        P0        PN     x_pos   x_width     y_pos   y_width    bg_pos  bg_width       fan      dpix       tth    number     DB_ID      File\n"  # noqa: E501
-    "#        1        15        10       167        12     163.5      72.9        55        24     False       194  0.00653668     42112         1  quicknxs-data/REF_M_42112.nxs.h5\n"  # noqa: E501
-    "# 0.183654        15        10     189.3        12     162.2      68.3        55        24     False       194  0.799577     42116           2  quicknxs-data/REF_M_42116.nxs.h5\n"  # noqa: E501
-    "#   0.1375        15        10     167.5        12     159.9      67.4        55        24     False       194  0.798876     42113           3  quicknxs-data/REF_M_42113.nxs.h5\n"  # noqa: E501
+    "#        1        15        10       167        12     163.5      72.9        55        24     False       194  0.00653668     42112         1  data/quicknxs-data/REF_M_42112.nxs.h5\n"  # noqa: E501
+    "# 0.183654        15        10     189.3        12     162.2      68.3        55        24     False       194  0.799577     42116           2  data/quicknxs-data/REF_M_42116.nxs.h5\n"  # noqa: E501
+    "#   0.1375        15        10     167.5        12     159.9      67.4        55        24     False       194  0.798876     42113           3  data/quicknxs-data/REF_M_42113.nxs.h5\n"  # noqa: E501
     "#\n"
     "# [Global Options]\n"
     "# name           value\n"
@@ -272,6 +271,22 @@ class TestDataManipulation(object):
 
         assert "# Mantid version 6.0.0" in result
         assert "Generated script" in result
+
+    @pytest.mark.datarepo
+    @pytest.mark.parametrize(
+        "data_file, expected",
+        [
+            ("REF_M_24945_event.nxs", {"mid_q": 6.26641549137356e-08, "is_direct_beam": False}),
+            ("REF_M_42112.nxs.h5", {"mid_q": 0.00013398646965655087, "is_direct_beam": False}),
+        ],
+    )
+    def test_extract_meta_data(self, data_server, data_file, expected):
+        """Test the extract_meta_data function"""
+        manager = DataManager(data_server.directory)
+        fp = f"{data_server.directory}/quicknxs-data/{data_file}"
+        metadata = extract_meta_data(fp)
+        assert metadata.mid_q == expected["mid_q"]
+        assert metadata.is_direct_beam == expected["is_direct_beam"]
 
 
 if __name__ == "__main__":