+++ test/unit/quicknxs/interfaces/data_handling/test_instrument.py
@@ -11,6 +11,17 @@ from quicknxs.interfaces.data_handling.instrument import mantid_algorithm_exec
 
 
 @pytest.mark.datarepo
+def test_load_data_pre_epics(data_server):
+    """Test load data with pre-epics cross-sections"""
+    conf = Configuration()
+    file_path = data_server.path_to("REF_M_24945_event.nxs")
+    ws_list = conf.instrument.load_data(file_path, conf)
+    assert len(ws_list) == 1
+    for ws in ws_list:
+        assert ws.getNumberEvents() == 7880217
+
+
+@pytest.mark.datarepo
 def test_load_data_deadtime(data_server):
     """Test load data with and without dead-time correction"""
     conf = Configuration()
@@ -84,3 +95,22 @@ def test_load_data_nbr_events_min(data_server):
     conf.apply_deadtime = True
     ws_list = conf.instrument.load_data(file_path, conf)
     assert len(ws_list) == 2
+
+
+@pytest.mark.parametrize(
+    "apply_deadtime",
+    [
+        False,
+        True,
+    ],
+)
+@pytest.mark.datarepo
+def test_load_unpolarized_data(data_server, apply_deadtime):
+    """Test load unpolarized data with Polarizer = 0 and Analyzer = 0"""
+    conf = Configuration()
+    conf.apply_deadtime = apply_deadtime
+    file_path = data_server.path_to("REF_M_41889")
+    ws_list = conf.instrument.load_data(file_path, conf)
+    assert len(ws_list) == 1
+    run = ws_list[0].getRun()
+    assert "cross_section_id" in run